Issue #1859: Refactor all target video JS methods
