+++ app/src/gecko/java/org/mozilla/tv/firefox/ext/EngineView.kt
@@ -10,7 +10,7 @@ import android.webkit.ValueCallback
 import android.widget.FrameLayout
 import mozilla.components.concept.engine.EngineView
 import org.mozilla.geckoview.GeckoView
-import org.mozilla.tv.firefox.webrender.FocusedDOMElementCache
+import org.mozilla.tv.firefox.webrender.FocusvideedDOMElementCache
 import java.util.WeakHashMap
 
 // Extension methods on the EngineView class. This is used for additional features that are not part
